pyproject.toml
[project]
name = "agency-ai-orchestrator"
version = "0.1.1"
dependencies = [
    "fastapi==0.128.0",
    "uvicorn==0.40.0",
    "langgraph==1.0.7",
    "langchain-anthropic==1.3.1",
    "httpx==0.28.1",
    "pydantic==2.12.5",
    "python-dotenv==1.2.1",
    # Google Drive API dependencies
    "google-auth==2.27.0",
    "google-auth-oauthlib==1.2.0",
    "google-api-python-client==2.116.0",
    # Document processing
    "python-docx==1.1.0",
    "pypdf==4.3.1",
    "Pillow>=11.0.0",
    "reportlab>=4.0.0",
    "beautifulsoup4>=4.12.3",
    # Observability & Evaluation
    "langfuse>=3.0.0",
    "ragas"
]

[tool.setuptools]
packages = ["app", "app.agents", "app.services"]